+++ tests/testthat/test-split_funs.R
@@ -1,8 +1,9 @@
 context("Split Functions")
 
-test_that("remove_split_levels works as expected with default flags", {
+test_that("remove_split_levels works as expected with factor variables", {
   my_split_fun <- remove_split_levels(excl = "ASIAN")
   
+  stopifnot(is.factor(DM$RACE))
   l <- basic_table() %>% 
     split_cols_by("ARM") %>%
     split_rows_by("RACE", split_fun = my_split_fun) %>%
@@ -13,8 +14,23 @@ test_that("remove_split_levels works as expected with default flags", {
   expect_false("ASIAN" %in% row.names(tab))
 })
 
-test_that("remove_split_levels works as expected when dropping not appearing levels", {
-  my_split_fun <- remove_split_levels(excl = "ASIAN", drop_levels = TRUE)
+test_that("remove_split_levels works as expected with character variables", {
+  my_split_fun <- remove_split_levels(excl = "ASIAN")
+  
+  l <- basic_table() %>% 
+    split_cols_by("ARM") %>%
+    split_rows_by("RACE", split_fun = my_split_fun) %>%
+    summarize_row_groups()
+  
+  DM2 <- DM
+  DM2$RACE <- as.character(DM2$RACE)
+  tab <- build_table(l, DM2)
+  
+  expect_false("ASIAN" %in% row.names(tab))
+})
+
+test_that("drop_and_remove_levels works as expected when dropping not appearing levels", {
+  my_split_fun <- drop_and_remove_levels(excl = "ASIAN")
   
   l <- basic_table() %>% 
     split_cols_by("ARM") %>%
@@ -29,23 +45,20 @@ test_that("remove_split_levels works as expected when dropping not appearing lev
   )
 })
 
-test_that("remove_split_levels works as expected when dropping not appearing levels and keeping data order", {
-  my_split_fun <- remove_split_levels(excl = "ASIAN", drop_levels = TRUE, keep_order = TRUE)
+test_that("drop_and_remove_levels also works with character variables", {
+  my_split_fun <- drop_and_remove_levels(excl = "ASIAN")
   
   l <- basic_table() %>% 
     split_cols_by("ARM") %>%
     split_rows_by("RACE", split_fun = my_split_fun) %>%
     summarize_row_groups()
   
-  DM <- DM[order(DM$RACE, decreasing = TRUE), ]
-  tab <- build_table(l, DM)
+  DM2 <- DM
+  DM2$RACE <- as.character(DM2$RACE)
+  tab <- build_table(l, DM2)
   
-  expect_identical(
+  expect_setequal(
     row.names(tab),
     setdiff(unique(DM$RACE), "ASIAN")
   )
-})
-
-test_that("remove_split_levels does not allow keeping order without dropping levels", {
-  expect_error(remove_split_levels(excl = "ASIAN", drop_levels = FALSE, keep_order = TRUE))
 })